+++ test/clearingHouse/ClearingHouse.swap.test.ts
@@ -92,7 +92,7 @@ describe("ClearingHouse.swap", () => {
                 amount: parseEther("1"),
                 sqrtPriceLimitX96: 0,
             })
-            initOpenNotional = await exchange.getOpenNotional(bob.address, baseToken.address)
+            initOpenNotional = await exchange.getTotalOpenNotional(bob.address, baseToken.address)
         })
 
         it("openNotional++", async () => {
@@ -140,7 +140,7 @@ describe("ClearingHouse.swap", () => {
             })
 
             it("openNotionalAbs--", async () => {
-                const openNotional = await exchange.getOpenNotional(bob.address, baseToken.address)
+                const openNotional = await exchange.getTotalOpenNotional(bob.address, baseToken.address)
                 // expect openNotion are same signed
                 expect(openNotional.mul(initOpenNotional).gt(0))
                 expect(openNotional.abs().lt(initOpenNotional.abs())).be.true
@@ -204,7 +204,7 @@ describe("ClearingHouse.swap", () => {
                 amount: parseEther("250"),
                 sqrtPriceLimitX96: 0,
             })
-            initOpenNotional = await exchange.getOpenNotional(bob.address, baseToken.address)
+            initOpenNotional = await exchange.getTotalOpenNotional(bob.address, baseToken.address)
             posSizeBefore = await accountBalance.getTotalPositionSize(bob.address, baseToken.address)
         })
 
@@ -253,7 +253,7 @@ describe("ClearingHouse.swap", () => {
             })
 
             it("openNotional--", async () => {
-                const openNotional = await exchange.getOpenNotional(bob.address, baseToken.address)
+                const openNotional = await exchange.getTotalOpenNotional(bob.address, baseToken.address)
                 // expect openNotion are same signed
                 expect(openNotional.mul(initOpenNotional).gt(0))
                 expect(openNotional.abs().lt(initOpenNotional.abs())).be.true
@@ -291,7 +291,7 @@ describe("ClearingHouse.swap", () => {
 
             it("reverse open notional 's signed", async () => {
                 // 400 - 206.6686875002 = 193.3313124998
-                const openNotional = await exchange.getOpenNotional(bob.address, baseToken.address)
+                const openNotional = await exchange.getTotalOpenNotional(bob.address, baseToken.address)
                 expect(openNotional).eq(parseEther("193.331312499999999962"))
             })
         })